chore(server): add development server environment
The development environment for the solution utilises a nodejs server
for hosting the web files and utilises webpack for loading web content
dynamically.

React has been included as a dependency as it will be used for
building the front end of the chat application.

A custom port can be provided as an environment variable for the nodejs
server, this is achieved by creating a `PORT` environment variable with
the desired port assigned to it e.g. `export PORT=3005`

The development environment can be started by running `npm start` which
will start the server by default on port 3001. The port can be provided
at start like so `PORT=3005 npm start`.

The README.md file has been updated to provide instructions of
installation and usage of the development server.