refactor(ChatPanel): move sub components into separate classes
The two existing components of the chat panel have now been extracted
into their own components to ensure the composition that ReactJS
teaches is adhered to.

The ChatPanel test has been updated to not mock the children components
so they will render correctly and the tests will still pass.

The ChatMessage has a basic test added to ensure the button label is
correct.