feat(i18n): add i18n for displayed text
To ensure the text within the app can be changed independently of the
component template (for example, to change the displayed language),
a translation library has been included that now displays the text
on page based on a given translation key.

A locales directory has been added to contain the translation files
along with a simple helper class that loads the i18n library and
configures it for use, this DRYs up the components.

The translated data is passed into the components as prop data to ensure
it is easy to mock during testing.