feat(messaging): add ability to send a local message
Utilising the flux pattern, it is now possible to type a message into
the `ChatMessage` component, which tracks the state of every key press
into the `<textarea/>` (which could be useful later on if want to know
if the other person is typing).

When the `ChatMessage` component has the button clicked, it dispatches
the current state of the `<textarea/>` value (the typed in message) via
the `MessageAction` object.

The `MessagesActions` object in turn dispatches the received message via
the `MessageDispatcher`. The `MessageStore` can now hold the global
state of the chat app and if a message is sent, the `MessageStore` picks
up the dispatched message and holds the current sent chat messages.

The `ChatHistory` component subscribes to the `MessageStore` and listens
for any changes, when a change is made to the state of `MessageStore`
the `ChatHistory` component retrieves the latest object of all of the
messages and updates the chat history window.

The `ChatHistory` component now displays the history in an unordered
list to ensure an appropriate structure.